+++ jarvis/executors/vpn_server.py
@@ -1,4 +1,5 @@
 import os
+from collections.abc import Generator
 from multiprocessing import Process
 from threading import Thread
 
@@ -9,9 +10,43 @@ from jarvis.modules.database import database
 from jarvis.modules.logger import config
 from jarvis.modules.logger.custom_logger import logger
 from jarvis.modules.models import models
-from jarvis.modules.utils import support
+from jarvis.modules.utils import support, util
 
 db = database.Database(database=models.fileio.base_db)
+
+
+def regional_phrase(phrase: str) -> Generator[str]:
+    """Converts alphabetical numbers into actual numbers.
+
+    Args:
+        phrase: Takes the phrase spoken as an argument.
+
+    Yields:
+        str:
+        Yields the word.
+    """
+    for word in phrase.split():
+        if num := util.words_to_number(input_=word):
+            yield str(num)
+        else:
+            yield word
+
+
+def extract_custom_region(phrase: str) -> str:
+    """Tries to find region name in the phrase.
+
+    Args:
+        phrase: Takes the phrase spoken as an argument.
+
+    Returns:
+        str:
+        Region name if a match is found.
+    """
+    phrase = " ".join(regional_phrase(phrase=phrase))
+    for region in vpn.settings.available_regions:
+        if region.replace('-', ' ') in phrase:
+            logger.info(f"Custom region chosen: {region}")
+            return region
 
 
 def vpn_server(phrase: str) -> None:
@@ -30,11 +65,16 @@ def vpn_server(phrase: str) -> None:
 
     phrase = phrase.lower()
     if 'start' in phrase or 'trigger' in phrase or 'initiate' in phrase or 'enable' in phrase or 'spin up' in phrase:
-        process = Process(target=vpn_server_switch, kwargs={'operation': 'enabled'})
-        speaker.speak(text=f'VPN Server has been initiated {models.env.title}! '
-                           'Login details will be sent to you shortly.')
+        if custom_region := extract_custom_region(phrase=phrase):
+            process = Process(target=vpn_server_switch, kwargs={'operation': 'enabled', 'custom_region': custom_region})
+            text = f'VPN Server has been initiated in {custom_region} {models.env.title}! ' \
+                   'Login details will be sent to you shortly.'
+        else:
+            process = Process(target=vpn_server_switch, kwargs={'operation': 'enabled'})
+            text = f'VPN Server has been initiated {models.env.title}! Login details will be sent to you shortly.'
+        speaker.speak(text=text)
     elif 'stop' in phrase or 'shut' in phrase or 'close' in phrase or 'disable' in phrase:
-        if not os.path.isfile('vpn_info.json'):
+        if not os.path.isfile(vpn.INFO_FILE):
             speaker.speak(text=f'Input file for VPN Server is missing {models.env.title}! '
                                'The VPN Server might have been shut down already.')
             return
@@ -49,22 +89,26 @@ def vpn_server(phrase: str) -> None:
     process.start()
 
 
-def vpn_server_switch(operation: str) -> None:
+def vpn_server_switch(operation: str, custom_region: str = None) -> None:
     """Automator to ``create`` or ``destroy`` a VPN server.
 
     Args:
-        operation: Takes ``enabled`` or ``disabled`` as an argument.
+        operation: Takes ``enabled`` or ``disabled`` as argument.
+        custom_region: Takes a custom AWS region as argument.
 
     See Also:
         - Check Read Me in `vpn-server <https://git.io/JzCbi>`__ for more information.
     """
     config.multiprocessing_logger(filename=os.path.join('logs', 'vpn_server_%d_%m_%Y_%H_%M.log'))
-    vpn_object = vpn.VPNServer(vpn_username=models.env.vpn_username or models.env.root_user,
-                               vpn_password=models.env.vpn_password or models.env.root_password,
-                               domain=models.env.vpn_domain, record_name=models.env.vpn_record_name,
-                               gmail_user=models.env.alt_gmail_user, gmail_pass=models.env.alt_gmail_pass,
-                               recipient=models.env.recipient or models.env.gmail_user,
-                               phone=models.env.phone_number, logger=logger)
+    kwargs = dict(vpn_username=models.env.vpn_username or models.env.root_user,
+                  vpn_password=models.env.vpn_password or models.env.root_password,
+                  domain=models.env.vpn_domain, record_name=models.env.vpn_record_name,
+                  gmail_user=models.env.alt_gmail_user, gmail_pass=models.env.alt_gmail_pass,
+                  recipient=models.env.recipient or models.env.gmail_user,
+                  phone=models.env.phone_number, logger=logger)
+    if custom_region:
+        kwargs['aws_region_name'] = custom_region
+    vpn_object = vpn.VPNServer(**kwargs)
     with db.connection:
         cursor = db.connection.cursor()
         cursor.execute("INSERT or REPLACE INTO vpn (state) VALUES (?);", (operation,))